refactor(ffmpeg): relocate watchdog into infra/ffmpeg, drop the orphaned tree
The shared FFmpeg process watchdog lived under internal/media/ffmpeg/watchdog —
an incomplete-migration artifact: internal/media/ffmpeg/ otherwise held nothing,
yet watchdog is imported by both infra/ffmpeg (runner) and infra/media/ffmpeg
(adapter). Move it into the shared-primitives tree (internal/infra/ffmpeg/watchdog)
and delete the orphaned parent.

Three ffmpeg trees -> two: a clean layered split remains (infra/ffmpeg = shared
primitives + VOD command builder + watchdog; infra/media/ffmpeg = live pipeline).
Pure git mv + import-path update; full build + ffmpeg/watchdog tests green,
golangci-lint --new-from-rev = 0 new issues.
